### What I did

loading plugins is the main cause of a slow pytest, come to learn
Ape's wasn't helping either.
This makes Ape not th

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

Before:

```
pytest --help -s  4.31s user 0.78s system 119% cpu 4.240 total
```

After:

```
pytest --help -s  2.08s user 0.30s system 97% cpu 2.439 total
```

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated